Review and update Validations doc
PAUSED: Jaokim and team are debatting removing or reworking error111.

Re-organize and re-section.

Add some info on exit codes.

Add a screenshot to show an example of `ws get:messages`.

Update Warning 207 screenshot to what it should be now: Error 111.

Turf note on Warning 207 and it only be applied to check command, it's
not true anymore.

Edits for clarity.

Todo:
- find references to old anchors and rename